# zookeeper-rest-api/src/main/java/com/rfranco/zookeeperrestapi/autogenerated/dtos/NodeCreationRequest.java
package com.rfranco.zookeeperrestapi.autogenerated.dtos;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import javax.validation.Valid;
import javax.validation.constraints.*;

/**
 * NodeCreationRequest
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2017-12-05T13:08:55.108+01:00")

public class NodeCreationRequest   {
  @JsonProperty("name")
  private String name = null;

  @JsonProperty("data")
  private String data = null;

  public NodeCreationRequest name(String name) {
    this.name = name;
    return this;
  }

   /**
   * The name of the new node.
   * @return name
  **/
  @ApiModelProperty(required = true, value = "The name of the new node.")
  @NotNull


  public String getName() {
    return name;
  }

  public void setName(String name) {
    this.name = name;
  }

  public NodeCreationRequest data(String data) {
    this.data = data;
    return this;
  }

   /**
   * The string data of the node.
   * @return data
  **/
  @ApiModelProperty(value = "The string data of the node.")


  public String getData() {
    return data;
  }

  public void setData(String data) {
    this.data = data;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    NodeCreationRequest nodeCreationRequest = (NodeCreationRequest) o;
    return Objects.equals(this.name, nodeCreationRequest.name) &&
        Objects.equals(this.data, nodeCreationRequest.data);
  }

  @Override
  public int hashCode() {
    return Objects.hash(name, data);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class NodeCreationRequest {\n");
    
    sb.append("    name: ").append(toIndentedString(name)).append("\n");
    sb.append("    data: ").append(toIndentedString(data)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}